recreate createEntity and use it in the entitycreate command
