+++ src/editor/lib/commands/EntityCreateCommand.js
@@ -1,8 +1,6 @@
 import Events from '../Events';
 import { Command } from '../command.js';
-import { createUniqueId } from '../entity.js';
-
-const NOT_COMPONENTS = ['id', 'class', 'mixin'];
+import { createEntity } from '../entity.js';
 
 /**
  * Helper function to add a new entity with a list of components
@@ -22,51 +20,13 @@ export class EntityCreateCommand extends Command {
 
   execute() {
     const definition = this.definition;
-    this.entity = document.createElement(definition.element || 'a-entity');
-    const entity = this.entity;
-
-    // Set id
-    if (definition.id) {
-      entity.id = definition.id;
-    } else {
-      this.entity.id = createUniqueId();
-    }
-
-    // Set class, mixin
-    for (const attribute of NOT_COMPONENTS) {
-      if (attribute !== 'id' && definition[attribute]) {
-        entity.setAttribute(attribute, definition[attribute]);
-      }
-    }
-
-    // Set data attributes
-    for (const key in definition) {
-      if (key.startsWith('data-')) {
-        entity.setAttribute(key, definition[key]);
-      }
-    }
-
-    // Set components
-    for (const componentName in definition.components) {
-      const componentValue = definition.components[componentName];
-      entity.setAttribute(componentName, componentValue);
-    }
-
-    // Emit event after entity is loaded
-    this.entity.addEventListener(
-      'loaded',
-      () => {
-        this.editor.selectEntity(this.entity);
-        Events.emit('entitycreated', this.entity);
-      },
-      { once: true }
-    );
-
-    // Add to parentEl if defined of fallback to scene container
+    const callback = (entity) => {
+      this.editor.selectEntity(entity);
+    };
     const parentEl =
       this.definition.parentEl ?? document.querySelector('#street-container');
-    parentEl.appendChild(this.entity);
-    return entity;
+    this.entity = createEntity(definition, callback, parentEl);
+    return this.entity;
   }
 
   undo() {
